requirements.txt
MetaTrader5>=5.0.45
python-telegram-bot>=21.0
python-dotenv>=1.0.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
schedule>=1.2.0
